resolve comments & restructure vc_di models
